refactor: extraer scoring de encuestas a src/utils/surveyScoring.js
